added some features, soveld some todos, checking if mysql user should be created, checking if should ouptut generated passwords for mysql

+++ add_web_user.sh
@@ -2,6 +2,24 @@
 # author: Aivars Lauzis
 # email: lauzis@inbox.lv
 
+
+
+CURR_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+SETTINGS_FILE=$CURR_DIR"/settings.cfg";
+
+#include settings
+#checks if settings file exists if not exits
+if [ -f "$SETTINGS_FILE" ]
+then
+	source $SETTINGS_FILE
+else
+	echo ERR:1 Settings file $SETTINGS_FILE not found... sorry, can not continue...
+	exit
+fi
+
+
+
+
 #todo check if all necassary toolas are there
 #todo do we have openssl
 #todo do we have mkpasswd
@@ -15,9 +33,9 @@
 
 #setting the username and php user name
 #todo from input not hardcoded
-NEW_USER_NAME="dev_uwp"
-PHP_USER_NAME="php_"$NEW_USER_NAME
-CURR_DIR=$('pwd')
+NEW_USER_NAME="uwp"
+
+
 MYSQL_USER_NAME=$NEW_USER_NAME
 
 
@@ -44,33 +62,43 @@ chown $NEW_USER_NAME /home/$NEW_USER_NAME -R
 chgrp $NEW_USER_NAME /home/$NEW_USER_NAME -R
 chgrp www-data /home/$NEW_USER_NAME/logs -R
 
-#creates php user, wihtout password
-groupadd $PHP_USER_NAMER
-sudo useradd -g php_$NEW_USER_NAME php_$NEW_USER_NAME
+CREATE_PHP_USER=0
+if [ $CREATE_PHP_USER ]
+then
+    #creates php user, wihtout password
+    PHP_USER_NAME="php_"$NEW_USER_NAME
+    groupadd $PHP_USER_NAME
+    sudo useradd -g $PHP_USER_NAME $PHP_USER_NAME
+
+fi
 
 #todo copy standart php.conf
 #todo copy standart nginx.conf
 
+CREATE_MYSQL_USER_AND_DATABASE=0
+if [ $CREATE_MYSQL_USER_AND_DATABASE ]
+then
+    #generate mysql users passowrd
+    MYSQL_USER_PASS="$(openssl rand -base64 20)"
 
-#todo chekc setting if wee need to create database
-#generate mysql users passowrd
-MYSQL_USER_PASS="$(openssl rand -base64 20)"
+    tmp_file=$CURR_DIR"/templates/tmp.sql"
+    yes | cp -rf $CURR_DIR"/templates/sql_user_and_database_setup.sql" $tmp_file
+    sed -i -e 's/{{MYSQL_USER}}/'$MYSQL_USER_NAME'/g' $tmp_file
+    sed -i -e 's/{{MYSQL_DB}}/'$MYSQL_USER_NAME'/g' $tmp_file
+    sed -i -e 's/{{MYSQL_PASSWORD}}/'MYSQL_USER_PASS'/g' $tmp_file
 
-tmp_file=$CURR_DIR"/templates/tmp.sql"
-yes | cp -rf $CURR_DIR"/templates/sql_user_and_database_setup.sql" $tmp_file
-sed -i -e 's/{{MYSQL_USER}}/'$MYSQL_USER_NAME'/g' $tmp_file
-sed -i -e 's/{{MYSQL_DB}}/'$MYSQL_USER_NAME'/g' $tmp_file
-sed -i -e 's/{{MYSQL_PASSWORD}}/'MYSQL_USER_PASS'/g' $tmp_file
+    mysql -u root -p < $tmp_file
+    #removes temporary sql file
+    rm $tmp_file
+fi
 
-
-#todo copy sql tamplate with replaced values
-
-mysql -u root -p < $tmp_file
-rm $tmp_file
 
 echo "U:"$NEW_USER_NAME
 echo "P:"$NEW_USER_PASS
 
-echo "MYSQL  U:"$NEW_USER_NAME
-echo "MYSQL  O:"$MYSQL_USER_PASS
-echo "MYSQL DB:"$NEW_USER_NAME
+if [ $CREATE_MYSQL_USER_AND_DATABASE ]
+then
+    echo "MYSQL  U:"$NEW_USER_NAME
+    echo "MYSQL  O:"$MYSQL_USER_PASS
+    echo "MYSQL DB:"$NEW_USER_NAME
+fi
